🚀 V. 1.1.1

+++ src/Settings.tsx
@@ -22,8 +22,6 @@ import InformationLineIcon from "remixicon-react/InformationLineIcon";
 import FileUploadLineIcon from "remixicon-react/FileUploadLineIcon";
 import FileDownloadLineIcon from "remixicon-react/FileDownloadLineIcon";
 import { useSwipeable } from "react-swipeable";
-import { Share } from "@capacitor/share";
-import JSZip from "jszip";
 
 async function createNotesDirectory() {
   const directoryPath = "notes";
@@ -247,6 +245,7 @@ const Settings: React.FC = () => {
       const currentDate = new Date();
       const formattedDate = currentDate.toISOString().split("T")[0]; // Format as YYYY-MM-DD
 
+      // Create the parent export folder
       const parentExportFolderPath = `export`;
       await Filesystem.mkdir({
         path: parentExportFolderPath,
@@ -254,21 +253,22 @@ const Settings: React.FC = () => {
         recursive: true,
       });
 
+      // Create the export folder structure
       const exportFolderName = `Beaver Notes ${formattedDate}`;
       const exportFolderPath = `${parentExportFolderPath}/${exportFolderName}`;
 
+      // Create the export folder
       await Filesystem.mkdir({
         path: exportFolderPath,
         directory: Directory.Documents,
         recursive: true,
       });
 
-      const exportedData: any = {
+      // Export data.json
+      let exportedData: any = {
         data: {
           notes: {},
-          lockedNotes: {},
         },
-        labels: [],
       };
 
       Object.values(notesState).forEach((note) => {
@@ -286,22 +286,33 @@ const Settings: React.FC = () => {
           updatedAt: updatedAtTimestamp,
           isBookmarked: note.isBookmarked,
           isArchived: note.isArchived,
-          isLocked: note.isLocked,
           lastCursorPosition: note.lastCursorPosition,
         };
-
-        exportedData.labels = exportedData.labels.concat(note.labels);
-
-        if (note.isLocked) {
-          exportedData.data.lockedNotes[note.id] = true;
-        }
       });
 
-      exportedData.labels = Array.from(new Set(exportedData.labels));
+      let jsonData = JSON.stringify(exportedData, null, 2);
 
-      const jsonData = JSON.stringify(exportedData, null, 2);
+      // Check if password protection is enabled
+      if (withPassword) {
+        // Prompt the user for a password
+        const passwordPrompt =
+          translations.settings.Inputpassword || "Enter password for export:";
+        const password = prompt(passwordPrompt);
+
+        // Check if the user provided a password
+        if (password !== null) {
+          // Encrypt the data using CryptoJS and the user's password
+          jsonData = CryptoJS.AES.encrypt(jsonData, password).toString();
+        } else {
+          // User canceled password input, abort export
+          console.log("Export canceled.");
+          return;
+        }
+      }
+
       const jsonFilePath = `${exportFolderPath}/data.json`;
 
+      // Save data.json
       await Filesystem.writeFile({
         path: jsonFilePath,
         data: jsonData,
@@ -309,6 +320,7 @@ const Settings: React.FC = () => {
         encoding: FilesystemEncoding.UTF8,
       });
 
+      // Check if the images folder exists
       const imagesFolderPath = `images`;
       let imagesFolderExists = false;
 
@@ -323,14 +335,17 @@ const Settings: React.FC = () => {
       }
 
       if (imagesFolderExists) {
+        // Export images folder
         const exportImagesFolderPath = `${exportFolderPath}/${imagesFolderPath}`;
 
+        // Create the images folder in the export directory
         await Filesystem.mkdir({
           path: exportImagesFolderPath,
           directory: Directory.Documents,
           recursive: true,
         });
 
+        // Copy images folder to export folder
         await Filesystem.copy({
           from: imagesFolderPath,
           to: exportImagesFolderPath,
@@ -338,63 +353,9 @@ const Settings: React.FC = () => {
         });
       }
 
-      const zip = new JSZip();
-      const exportFolderZip = zip.folder(`Beaver Notes ${formattedDate}`);
-
-      const exportFolderFiles = await Filesystem.readdir({
-        path: exportFolderPath,
-        directory: Directory.Documents,
-      });
-
-      await Promise.all(
-        exportFolderFiles.files.map(async (file) => {
-          const filePath = `${exportFolderPath}/${file.name}`;
-          const fileContent = await Filesystem.readFile({
-            path: filePath,
-            directory: Directory.Documents,
-            encoding: FilesystemEncoding.UTF8,
-          });
-          exportFolderZip!.file(file.name, fileContent.data);
-        })
-      );
-
-      const zipContentBase64 = await zip.generateAsync({ type: "base64" });
-
-      const zipFilePath = `/Beaver_Notes_${formattedDate}.zip`;
-      await Filesystem.writeFile({
-        path: zipFilePath,
-        data: zipContentBase64,
-        directory: Directory.Cache,
-      });
-
-      await shareZipFile();
-
       alert(translations.home.exportSuccess);
     } catch (error) {
       alert(translations.home.exportError + (error as any).message);
-    }
-  };
-
-  const shareZipFile = async () => {
-    try {
-      const currentDate = new Date();
-      const formattedDate = currentDate.toISOString().split("T")[0]; // Format as YYYY-MM-DD
-      const zipFilePath = `/Beaver_Notes_${formattedDate}.zip`;
-
-      const result = await Filesystem.getUri({
-        directory: Directory.Cache,
-        path: zipFilePath,
-      });
-
-      const resolvedFilePath = result.uri;
-
-      await Share.share({
-        title: `${translations.home.shareTitle}`,
-        url: resolvedFilePath,
-        dialogTitle: `${translations.home.shareTitle}`,
-      });
-    } catch (error) {
-      alert(translations.home.shareError + (error as any).message);
     }
   };
